+++ test/enum-collection.test.ts
@@ -1,4 +1,4 @@
-import { Enum, KEYS, VALUES } from '../src';
+import { Enum, ITEMS, KEYS, VALUES } from '../src';
 import { EnumValuesArray } from '../src/enum-values';
 import { locales, sillyLocalize, StandardWeekConfig } from './data/week-config';
 import { getStandardWeekData } from './data/week-data';
@@ -15,24 +15,24 @@ describe('the EnumCollectionClass api', () => {
 
   test('[values] should be able to return an array of enum items', () => {
     const week = Enum(StandardWeekConfig);
-    expect(Array.isArray(week.values)).toBeTruthy();
-    expect(toPlainEnums(week.values)).toEqual(getStandardWeekData(locales));
+    expect(Array.isArray(week.items)).toBeTruthy();
+    expect(toPlainEnums(week.items)).toEqual(getStandardWeekData(locales));
   });
 
   test('api methods should be same as EnumValuesArray', () => {
     const week = Enum(StandardWeekConfig);
-    expect(week.label(1)).toBe(week.values.label(1));
-    expect(week.key(1)).toBe(week.values.key(1));
-    expect(week.has(1)).toBe(week.values.has(1));
-    expect(week.raw()).toBe(week.values.raw());
-    expect(week.toSelect()).toEqual(week.values.toSelect());
-    expect(week.options()).toEqual(week.values.options());
-    expect(week.toMenu()).toEqual(week.values.toMenu());
-    expect(week.menus()).toEqual(week.values.menus());
-    expect(week.toFilter()).toEqual(week.values.toFilter());
-    expect(week.filters()).toEqual(week.values.filters());
-    expect(week.toValueMap()).toEqual(week.values.toValueMap());
-    expect(week.valuesEnum()).toEqual(week.values.valuesEnum());
+    expect(week.label(1)).toBe(week.items.label(1));
+    expect(week.key(1)).toBe(week.items.key(1));
+    expect(week.has(1)).toBe(week.items.has(1));
+    expect(week.raw()).toBe(week.items.raw());
+    expect(week.toSelect()).toEqual(week.items.toSelect());
+    expect(week.options()).toEqual(week.items.options());
+    expect(week.toMenu()).toEqual(week.items.toMenu());
+    expect(week.menus()).toEqual(week.items.menus());
+    expect(week.toFilter()).toEqual(week.items.toFilter());
+    expect(week.filters()).toEqual(week.items.filters());
+    expect(week.toValueMap()).toEqual(week.items.toValueMap());
+    expect(week.valuesEnum()).toEqual(week.items.valuesEnum());
   });
 
   test('[keys,values] should be able to auto renaming to a non-conflicting name', () => {
@@ -46,39 +46,42 @@ describe('the EnumCollectionClass api', () => {
       toMenu: { value: 7, label: 'toMenu' },
       raw: { value: 99, label: 'raw' },
       keys: { value: 101, label: 'foo' },
-    };
+    } as const;
     const strangeEnum = Enum(strangeEnumConfig);
     expect(strangeEnum.label).toBe(1);
-    expect(strangeEnum.values.label(1)).toBe('label');
+    expect(strangeEnum.items.label(1)).toBe('label');
     expect(strangeEnum.key).toBe(2);
-    expect(strangeEnum.values.key(2)).toBe('key');
+    expect(strangeEnum.items.key(2)).toBe('key');
     expect(strangeEnum.has).toBe(3);
-    expect(strangeEnum.values.has(3)).toBe(true);
+    expect(strangeEnum.items.has(3)).toBe(true);
     expect(strangeEnum.toSelect).toBe(4);
-    expect(strangeEnum.values.toSelect()).toHaveLength(Object.keys(strangeEnumConfig).length);
+    expect(strangeEnum.items.toSelect()).toHaveLength(Object.keys(strangeEnumConfig).length);
     expect(strangeEnum.toFilter).toBe(6);
-    expect(strangeEnum.values.toFilter()).toHaveLength(Object.keys(strangeEnumConfig).length);
+    expect(strangeEnum.items.toFilter()).toHaveLength(Object.keys(strangeEnumConfig).length);
     expect(strangeEnum.toMenu).toBe(7);
-    expect(strangeEnum.values.toMenu()).toHaveLength(Object.keys(strangeEnumConfig).length);
+    expect(strangeEnum.items.toMenu()).toHaveLength(Object.keys(strangeEnumConfig).length);
     expect(strangeEnum.toValueMap).toBe(5);
-    expect(Object.keys(strangeEnum.values.toValueMap())).toEqual(
+    expect(Object.keys(strangeEnum.items.toValueMap())).toEqual(
       Object.keys(strangeEnumConfig).map((key) =>
         strangeEnumConfig[key as keyof typeof strangeEnumConfig].value.toString()
       )
     );
     expect(strangeEnum.keys).toBe(101);
-    expect(strangeEnum.values.map((i) => i.key)).toEqual(Object.keys(strangeEnumConfig));
+    expect(strangeEnum.items.map((i) => i.key)).toEqual(Object.keys(strangeEnumConfig));
 
     const strangerEnumConfig = {
       ...strangeEnumConfig,
-      values: { value: 102, label: 'bar' },
-    };
+      items: { value: 101, label: 'bar' },
+      values: { value: 102, label: 'baz' },
+    } as const;
     const strangerEnum = Enum(strangerEnumConfig);
+    expect(strangerEnum.items).toBe(101);
     expect(strangerEnum.values).toBe(102);
     expect(strangerEnum[KEYS]).toEqual(Object.keys(strangerEnumConfig));
+    expect(strangerEnum[ITEMS]).toBeInstanceOf(EnumValuesArray);
     expect(strangerEnum[VALUES]).toBeInstanceOf(EnumValuesArray);
-    expect(Array.isArray(strangerEnum[VALUES])).toBeTruthy();
-    expect(toPlainEnums(strangerEnum[VALUES])).toEqual(
+    expect(Array.isArray(strangerEnum[ITEMS])).toBeTruthy();
+    expect(toPlainEnums(strangerEnum[ITEMS])).toEqual(
       Object.keys(strangerEnumConfig).map((key) => ({
         key,
         value: strangerEnumConfig[key as keyof typeof strangerEnumConfig].value,